move v4 to v3
