--- web_service_v4/__init__.py
@@ -1,2 +1,0 @@
-"""Web service v4."""
-
